Move the copy-from-array code into DataArrayUtilities for reusability
